requirements.txt
argbind>=0.3.7
descript-audiotools>=0.7.2
einops
numpy==1.26
tqdm
numba>=0.5.7
tensorboard
torchmetrics
